Fix UTscapy timeout logs
